fix: resolve ESLint configuration for mock files
- Add ESLint rules to allow common patterns in mock files
- Fix require() imports in test files by using ES6 module imports
- Remove unused eslint-disable directives

This ensures npm run lint passes without errors, resolving the main
ESLint configuration issues blocking CI/CD.

Resolves #246
