+++ apps/web/src/__tests__/__mocks__/@/components/ui/label.tsx
@@ -4,7 +4,6 @@ import { cn } from '@/lib/utils';
 export type LabelProps = React.LabelHTMLAttributes<HTMLLabelElement>;
 
 const Label = React.forwardRef<HTMLLabelElement, LabelProps>(({ className, ...props }, ref) => (
-  // eslint-disable-next-line jsx-a11y/label-has-associated-control
   <label
     ref={ref}
     className={cn(
