+++ apps/web/src/__tests__/__mocks__/@radix-ui/react-checkbox.js
@@ -1,5 +1,4 @@
 // Mock for @radix-ui/react-checkbox
-// eslint-disable-next-line @typescript-eslint/no-require-imports
 const React = require('react');
 
 // Checkbox component
