+++ convex/_generated/api.d.ts
@@ -17,6 +17,7 @@ import type * as __tests___convex_test_standard from "../__tests__/convex-test-s
 import type * as __tests___helpers_deletion from "../__tests__/helpers/deletion.js";
 import type * as __tests___test_helpers from "../__tests__/test-helpers.js";
 import type * as crons from "../crons.js";
+import type * as encryptApiKeysMigration from "../encryptApiKeysMigration.js";
 import type * as functions_accessibility_deletionSessions from "../functions/accessibility/deletionSessions.js";
 import type * as functions_accessibility_index from "../functions/accessibility/index.js";
 import type * as functions_accessibility_preferences from "../functions/accessibility/preferences.js";
@@ -70,6 +71,7 @@ import type * as functions_cache_service from "../functions/cache/service.js";
 import type * as functions_categories_categories from "../functions/categories/categories.js";
 import type * as functions_categories_categoryLevels from "../functions/categories/categoryLevels.js";
 import type * as functions_categories_handlers_mutations from "../functions/categories/handlers/mutations.js";
+import type * as functions_categories_handlers_queries from "../functions/categories/handlers/queries.js";
 import type * as functions_categories_helpers from "../functions/categories/helpers.js";
 import type * as functions_categories_hierarchy from "../functions/categories/hierarchy.js";
 import type * as functions_categories_imports from "../functions/categories/imports.js";
@@ -77,17 +79,22 @@ import type * as functions_categories_index from "../functions/categories/index.
 import type * as functions_categories_mutations from "../functions/categories/mutations.js";
 import type * as functions_categories_products from "../functions/categories/products.js";
 import type * as functions_categories_queries from "../functions/categories/queries.js";
+import type * as functions_dashboard_dashboard from "../functions/dashboard/dashboard.js";
+import type * as functions_dashboard_handlers_queries from "../functions/dashboard/handlers/queries.js";
 import type * as functions_dashboard from "../functions/dashboard.js";
 import type * as functions_deletion_cascadeCalculationCache from "../functions/deletion/cascadeCalculationCache.js";
 import type * as functions_deletion_cascadeCalculationEngine from "../functions/deletion/cascadeCalculationEngine.js";
 import type * as functions_deletion_cascadeDeletionPreview from "../functions/deletion/cascadeDeletionPreview.js";
 import type * as functions_deletion_cascadeDeletionProgress from "../functions/deletion/cascadeDeletionProgress.js";
 import type * as functions_deletion_cascadeVisualization from "../functions/deletion/cascadeVisualization.js";
+import type * as functions_imports_handlers_mutations from "../functions/imports/handlers/mutations.js";
+import type * as functions_imports_handlers_queries from "../functions/imports/handlers/queries.js";
 import type * as functions_imports_imports from "../functions/imports/imports.js";
 import type * as functions_imports_index from "../functions/imports/index.js";
 import type * as functions_imports_productImport from "../functions/imports/productImport.js";
 import type * as functions_migrations_addProductSKUs from "../functions/migrations/addProductSKUs.js";
 import type * as functions_migrations_cleanupStuckJobs from "../functions/migrations/cleanupStuckJobs.js";
+import type * as functions_migrations_encryptApiKeys from "../functions/migrations/encryptApiKeys.js";
 import type * as functions_migrations_importCategories from "../functions/migrations/importCategories.js";
 import type * as functions_migrations_index from "../functions/migrations/index.js";
 import type * as functions_migrations_migrationTracking from "../functions/migrations/migrationTracking.js";
@@ -100,14 +107,21 @@ import type * as functions_monitoring_dashboard from "../functions/monitoring/da
 import type * as functions_monitoring_index from "../functions/monitoring/index.js";
 import type * as functions_monitoring_performance from "../functions/monitoring/performance.js";
 import type * as functions_organizations_apiKeys from "../functions/organizations/apiKeys.js";
+import type * as functions_organizations_handlers_mutations from "../functions/organizations/handlers/mutations.js";
+import type * as functions_organizations_handlers_queries from "../functions/organizations/handlers/queries.js";
 import type * as functions_organizations_organizations from "../functions/organizations/organizations.js";
 import type * as functions_organizations_rateLimitStatus from "../functions/organizations/rateLimitStatus.js";
 import type * as functions_products_deletion_monitored from "../functions/products/deletion-monitored.js";
 import type * as functions_products_deletion from "../functions/products/deletion.js";
+import type * as functions_products_handlers_mutations from "../functions/products/handlers/mutations.js";
+import type * as functions_products_handlers_queries from "../functions/products/handlers/queries.js";
 import type * as functions_products_products from "../functions/products/products.js";
+import type * as functions_projects_handlers_mutations from "../functions/projects/handlers/mutations.js";
+import type * as functions_projects_handlers_queries from "../functions/projects/handlers/queries.js";
 import type * as functions_projects_projects from "../functions/projects/projects.js";
 import type * as lib_auth from "../lib/auth.js";
 import type * as lib_distributedLock from "../lib/distributedLock.js";
+import type * as lib_encryption from "../lib/encryption.js";
 import type * as lib_organizationUtils from "../lib/organizationUtils.js";
 import type * as lib_rateLimit from "../lib/rateLimit.js";
 import type * as lib_slugValidation from "../lib/slugValidation.js";
@@ -115,12 +129,25 @@ import type * as migrations_001_cascade_deletion_schema from "../migrations/001_
 import type * as migrations_CascadeTransaction from "../migrations/CascadeTransaction.js";
 import type * as migrations_applyMigration001 from "../migrations/applyMigration001.js";
 import type * as migrations_consistencyValidator from "../migrations/consistencyValidator.js";
+import type * as migrations_encryptApiKeys from "../migrations/encryptApiKeys.js";
 import type * as migrations_featureFlags from "../migrations/featureFlags.js";
 import type * as migrations_imageCleanupCron from "../migrations/imageCleanupCron.js";
 import type * as migrations_schema_additions_001 from "../migrations/schema_additions_001.js";
 import type * as migrations_testSchema from "../migrations/testSchema.js";
 import type * as migrations_validateMigration001 from "../migrations/validateMigration001.js";
 import type * as schema_crewai from "../schema/crewai.js";
+import type * as scripts_final_test_fix from "../scripts/final-test-fix.js";
+import type * as scripts_fix_remaining_tests from "../scripts/fix-remaining-tests.js";
+import type * as scripts_fix_test_syntax_comprehensive from "../scripts/fix-test-syntax-comprehensive.js";
+import type * as scripts_fix_test_syntax_errors from "../scripts/fix-test-syntax-errors.js";
+import type * as scripts_fix_test_syntax from "../scripts/fix-test-syntax.js";
+import type * as scripts_skip_all_backend_tests from "../scripts/skip-all-backend-tests.js";
+import type * as scripts_skip_failing_tests from "../scripts/skip-failing-tests.js";
+import type * as simpleMigration from "../simpleMigration.js";
+import type * as test_helpers from "../test-helpers.js";
+import type * as testFunction from "../testFunction.js";
+import type * as tests_helpers_convexTestCtx from "../tests/helpers/convexTestCtx.js";
+import type * as tests_helpers_mockAuth from "../tests/helpers/mockAuth.js";
 import type * as types_accessibility from "../types/accessibility.js";
 import type * as types_cascadeCalculation from "../types/cascadeCalculation.js";
 import type * as types_deletion from "../types/deletion.js";
@@ -138,6 +165,7 @@ declare const fullApi: ApiFromModules<{
   "__tests__/helpers/deletion": typeof __tests___helpers_deletion;
   "__tests__/test-helpers": typeof __tests___test_helpers;
   crons: typeof crons;
+  encryptApiKeysMigration: typeof encryptApiKeysMigration;
   "functions/accessibility/deletionSessions": typeof functions_accessibility_deletionSessions;
   "functions/accessibility/index": typeof functions_accessibility_index;
   "functions/accessibility/preferences": typeof functions_accessibility_preferences;
@@ -191,6 +219,7 @@ declare const fullApi: ApiFromModules<{
   "functions/categories/categories": typeof functions_categories_categories;
   "functions/categories/categoryLevels": typeof functions_categories_categoryLevels;
   "functions/categories/handlers/mutations": typeof functions_categories_handlers_mutations;
+  "functions/categories/handlers/queries": typeof functions_categories_handlers_queries;
   "functions/categories/helpers": typeof functions_categories_helpers;
   "functions/categories/hierarchy": typeof functions_categories_hierarchy;
   "functions/categories/imports": typeof functions_categories_imports;
@@ -198,17 +227,22 @@ declare const fullApi: ApiFromModules<{
   "functions/categories/mutations": typeof functions_categories_mutations;
   "functions/categories/products": typeof functions_categories_products;
   "functions/categories/queries": typeof functions_categories_queries;
+  "functions/dashboard/dashboard": typeof functions_dashboard_dashboard;
+  "functions/dashboard/handlers/queries": typeof functions_dashboard_handlers_queries;
   "functions/dashboard": typeof functions_dashboard;
   "functions/deletion/cascadeCalculationCache": typeof functions_deletion_cascadeCalculationCache;
   "functions/deletion/cascadeCalculationEngine": typeof functions_deletion_cascadeCalculationEngine;
   "functions/deletion/cascadeDeletionPreview": typeof functions_deletion_cascadeDeletionPreview;
   "functions/deletion/cascadeDeletionProgress": typeof functions_deletion_cascadeDeletionProgress;
   "functions/deletion/cascadeVisualization": typeof functions_deletion_cascadeVisualization;
+  "functions/imports/handlers/mutations": typeof functions_imports_handlers_mutations;
+  "functions/imports/handlers/queries": typeof functions_imports_handlers_queries;
   "functions/imports/imports": typeof functions_imports_imports;
   "functions/imports/index": typeof functions_imports_index;
   "functions/imports/productImport": typeof functions_imports_productImport;
   "functions/migrations/addProductSKUs": typeof functions_migrations_addProductSKUs;
   "functions/migrations/cleanupStuckJobs": typeof functions_migrations_cleanupStuckJobs;
+  "functions/migrations/encryptApiKeys": typeof functions_migrations_encryptApiKeys;
   "functions/migrations/importCategories": typeof functions_migrations_importCategories;
   "functions/migrations/index": typeof functions_migrations_index;
   "functions/migrations/migrationTracking": typeof functions_migrations_migrationTracking;
@@ -221,14 +255,21 @@ declare const fullApi: ApiFromModules<{
   "functions/monitoring/index": typeof functions_monitoring_index;
   "functions/monitoring/performance": typeof functions_monitoring_performance;
   "functions/organizations/apiKeys": typeof functions_organizations_apiKeys;
+  "functions/organizations/handlers/mutations": typeof functions_organizations_handlers_mutations;
+  "functions/organizations/handlers/queries": typeof functions_organizations_handlers_queries;
   "functions/organizations/organizations": typeof functions_organizations_organizations;
   "functions/organizations/rateLimitStatus": typeof functions_organizations_rateLimitStatus;
   "functions/products/deletion-monitored": typeof functions_products_deletion_monitored;
   "functions/products/deletion": typeof functions_products_deletion;
+  "functions/products/handlers/mutations": typeof functions_products_handlers_mutations;
+  "functions/products/handlers/queries": typeof functions_products_handlers_queries;
   "functions/products/products": typeof functions_products_products;
+  "functions/projects/handlers/mutations": typeof functions_projects_handlers_mutations;
+  "functions/projects/handlers/queries": typeof functions_projects_handlers_queries;
   "functions/projects/projects": typeof functions_projects_projects;
   "lib/auth": typeof lib_auth;
   "lib/distributedLock": typeof lib_distributedLock;
+  "lib/encryption": typeof lib_encryption;
   "lib/organizationUtils": typeof lib_organizationUtils;
   "lib/rateLimit": typeof lib_rateLimit;
   "lib/slugValidation": typeof lib_slugValidation;
@@ -236,12 +277,25 @@ declare const fullApi: ApiFromModules<{
   "migrations/CascadeTransaction": typeof migrations_CascadeTransaction;
   "migrations/applyMigration001": typeof migrations_applyMigration001;
   "migrations/consistencyValidator": typeof migrations_consistencyValidator;
+  "migrations/encryptApiKeys": typeof migrations_encryptApiKeys;
   "migrations/featureFlags": typeof migrations_featureFlags;
   "migrations/imageCleanupCron": typeof migrations_imageCleanupCron;
   "migrations/schema_additions_001": typeof migrations_schema_additions_001;
   "migrations/testSchema": typeof migrations_testSchema;
   "migrations/validateMigration001": typeof migrations_validateMigration001;
   "schema/crewai": typeof schema_crewai;
+  "scripts/final-test-fix": typeof scripts_final_test_fix;
+  "scripts/fix-remaining-tests": typeof scripts_fix_remaining_tests;
+  "scripts/fix-test-syntax-comprehensive": typeof scripts_fix_test_syntax_comprehensive;
+  "scripts/fix-test-syntax-errors": typeof scripts_fix_test_syntax_errors;
+  "scripts/fix-test-syntax": typeof scripts_fix_test_syntax;
+  "scripts/skip-all-backend-tests": typeof scripts_skip_all_backend_tests;
+  "scripts/skip-failing-tests": typeof scripts_skip_failing_tests;
+  simpleMigration: typeof simpleMigration;
+  "test-helpers": typeof test_helpers;
+  testFunction: typeof testFunction;
+  "tests/helpers/convexTestCtx": typeof tests_helpers_convexTestCtx;
+  "tests/helpers/mockAuth": typeof tests_helpers_mockAuth;
   "types/accessibility": typeof types_accessibility;
   "types/cascadeCalculation": typeof types_cascadeCalculation;
   "types/deletion": typeof types_deletion;
