+++ convex/functions/ai/monitoring/costAnalyzer.ts
@@ -1,4 +1,4 @@
-import { internal } from "../../_generated/api";
+import { internal } from "../../../_generated/api";
 /**
  * CrewAI Cost Analyzer
  *
